Task to cross-compile all Rubies
Instead of delaying the compilation of the cross-version of Ruby, we
trigger all the compilation at once.

- Use Ubuntu's GCC to compile 1.9.3 to match the GCC version used in
Windows (DevKit)

We do this to avoid cross-compilation of some gems detect
characteristics of mingw-w64 headers that will not be present on
Windows version of 1.9.3

This happen when using mingw-w64 to cross-compile projects like
EventMachine that uses some settings of mingw-w64 headers not present
in Windows counterpart for Ruby 1.9.3

Doing this we reduce the chances of messing things up when
cross-compiling gems.

- Turn debugging off for Ruby 2.0

For some reason, Ruby 2.0 enables it's debugging flags when
cross-compiled, resulting in massive binaries for all extensions
compiled with it.

While this has been reported to Ruby-Core developers, this still hasn't
been fixed yet.

Instead, reduce the debugging symbols to the minimum to allow gdb
remain usable.