Change Rubies installation to reduce duplication
Ruby 1.9.x and 2.x requires Ruby 1.8.x as pre-requisite.

When installing 2.1, RVM triggers the installation of 1.8.7-head
before, and then we install latest 1.8.7.

This change ensures that Ruby 1.8.7 is installed first so others
can depend on it and avoid multiple 1.8.7 installations.

Due Ruby 1.8.7 EOL, we are going to set a specific patchlevel for
it to avoid rvm installing `head` version of it too.