Use Ruby 1.9.3 now as base for cross-compilation
In the past, the only version that was capable of cross-compile
and target 1.8.7 was *just* 1.8.7.

But now, 1.8.7 is no longer capable of cross-compile to 2.1.x,
which put us in a weird situation for those gems that still need
or want to support 1.8.7

After some testing, seems that Ruby developers have patched the
issues around `mkmf` and backported into Ruby 1.9.3, which is now
capable of cross-compile to Ruby 1.8.7, yay!

Once again, regressions that are not regressions and happy endings
for those who still are forced to maintain compatibility with
Ruby 1.8.7.